revert: cancel file rename for now
