+++ src/utils/loading-animation.js
